refactor(reporting): extract the shared run/registry core to _shared.py
Task 2 of beautify-phase-2a: move the SQL-target constants, the
source/metric registry, the curated-report run pipeline (_prepare_run and
its collaborators), and the live-SQL sandbox's auth/audit helpers out of
nx_lib/views/reporting/__init__.py into a new _shared.py. __init__.py
re-exports every moved name under its original nx_lib.views.reporting.<name>
path, so nx_lib/reporting/runner.py's rv.<attr> reads and ai.py's
function-local 'from . import ...' pattern are both unaffected.

Retargeted ~13 unittest.mock.patch()/monkeypatch.setattr() call sites (in
tests/integration/test_reporting_metrics_routes.py,
tests/unit/test_reporting_registry_cache.py, and
tests/unit/test_reporting_sql_targets.py) from
nx_lib.views.reporting.<name> to nx_lib.views.reporting._shared.<name> --
these patch functions that are now called internally within _shared.py
(e.g. _prepare_run calling _get_effective_source, or _authorize_sql_target
calling has_permission), where the package-level patch no longer reaches
the call site. Verified live by temporarily breaking the real
_get_effective_source and confirming an unmocked test failed, then
reverting.

Added tests/unit/test_reporting_shared_exports.py: an explicit-list
contract test asserting the 12 attributes runner.py reads off
nx_lib.views.reporting all resolve.

Full reporting integration + unit suite green (733 passed) except the
known-unrelated locale-leak flake
(test_ai_agent_empty_answer_gets_artifact_aware_fallback). Smoke-tested
ops/run_scheduled_reports.py --dry-run against INT: ran clean end to end.
