+++ tests/integration/test_reporting_metrics_routes.py
@@ -207,19 +207,26 @@ _CATALOG = [
 
 
 def test_run_with_metric_returns_aggregated_rows(admin_client):
+    # _prepare_run and its collaborators (_get_effective_source, has_permission,
+    # _metrics_for_source, fetch_docprocessing_catalog, _allowed_processes,
+    # _load_process_configs, _load_field_col_maps, build_table_query) all live
+    # in nx_lib.views.reporting._shared now (beautify-phase-2a, Task 2) and
+    # call each other from there -- patching them on the package re-export
+    # would not reach these internal calls. _execute is still patched on the
+    # package because api_run() (in __init__.py) calls it directly.
     with (
-        patch("nx_lib.views.reporting._get_effective_source", return_value=_DOCPROC_SOURCE),
-        patch("nx_lib.views.reporting.has_permission", return_value=True),
+        patch("nx_lib.views.reporting._shared._get_effective_source", return_value=_DOCPROC_SOURCE),
+        patch("nx_lib.views.reporting._shared.has_permission", return_value=True),
         patch(
-            "nx_lib.views.reporting._metrics_for_source",
+            "nx_lib.views.reporting._shared._metrics_for_source",
             return_value={"doc_count": {"aggregation": "count", "base_field": None}},
         ),
-        patch("nx_lib.views.reporting.fetch_docprocessing_catalog", return_value=_CATALOG),
-        patch("nx_lib.views.reporting._allowed_processes", return_value=["acme.invoice"]),
-        patch("nx_lib.views.reporting._load_process_configs", return_value=[]),
-        patch("nx_lib.views.reporting._load_field_col_maps", return_value={}),
+        patch("nx_lib.views.reporting._shared.fetch_docprocessing_catalog", return_value=_CATALOG),
+        patch("nx_lib.views.reporting._shared._allowed_processes", return_value=["acme.invoice"]),
+        patch("nx_lib.views.reporting._shared._load_process_configs", return_value=[]),
+        patch("nx_lib.views.reporting._shared._load_field_col_maps", return_value={}),
         patch(
-            "nx_lib.views.reporting.build_table_query",
+            "nx_lib.views.reporting._shared.build_table_query",
             return_value=("SELECT [client], COUNT(*) AS [doc_count] FROM x GROUP BY [client]", []),
         ),
         patch("nx_lib.views.reporting._execute", return_value=[["Acme", 30]]),
@@ -236,14 +243,14 @@ def test_run_with_metric_returns_aggregated_rows(admin_client):
 def test_run_with_unknown_metric_returns_400(admin_client):
     bad = {**_RUN_DEF, "metrics": [{"metric": "ghost_metric"}]}
     with (
-        patch("nx_lib.views.reporting._get_effective_source", return_value=_DOCPROC_SOURCE),
-        patch("nx_lib.views.reporting.has_permission", return_value=True),
+        patch("nx_lib.views.reporting._shared._get_effective_source", return_value=_DOCPROC_SOURCE),
+        patch("nx_lib.views.reporting._shared.has_permission", return_value=True),
         patch(
-            "nx_lib.views.reporting._metrics_for_source",
+            "nx_lib.views.reporting._shared._metrics_for_source",
             return_value={"doc_count": {"aggregation": "count", "base_field": None}},
         ),
-        patch("nx_lib.views.reporting.fetch_docprocessing_catalog", return_value=_CATALOG),
-        patch("nx_lib.views.reporting._allowed_processes", return_value=["acme.invoice"]),
+        patch("nx_lib.views.reporting._shared.fetch_docprocessing_catalog", return_value=_CATALOG),
+        patch("nx_lib.views.reporting._shared._allowed_processes", return_value=["acme.invoice"]),
     ):
         resp = admin_client.post("/api/reporting/run", json=bad)
     # validate_report_definition rejects an unknown metric code (not in metric_codes).
