+++ tests/unit/test_reporting_sql_targets.py
@@ -9,28 +9,34 @@ it is monkeypatched here — no TEST access profile holds SQL-run-without-octopu
 import pytest
 
 from nx_lib.views import reporting
+from nx_lib.views.reporting import _shared
+
+# `_authorize_sql_target` and its internal `has_permission` call both live in
+# _shared.py now (beautify-phase-2a, Task 2) and call each other from there --
+# monkeypatching has_permission on the package re-export would not reach that
+# internal call, so it is patched on `_shared` directly.
 
 
 def test_statistics_target_allowed_with_base_perm(monkeypatch):
-    monkeypatch.setattr(reporting, "has_permission", lambda p: p == "reporting.sql.run")
+    monkeypatch.setattr(_shared, "has_permission", lambda p: p == "reporting.sql.run")
     reporting._authorize_sql_target("statistics")  # no raise
 
 
 def test_octopus_target_denied_without_target_perm(monkeypatch):
     # Has the base SQL-run perm but not the Octopus target perm.
-    monkeypatch.setattr(reporting, "has_permission", lambda p: p == "reporting.sql.run")
+    monkeypatch.setattr(_shared, "has_permission", lambda p: p == "reporting.sql.run")
     with pytest.raises(PermissionError):
         reporting._authorize_sql_target("octopus")
 
 
 def test_octopus_target_allowed_with_target_perm(monkeypatch):
-    monkeypatch.setattr(reporting, "has_permission", lambda p: True)
+    monkeypatch.setattr(_shared, "has_permission", lambda p: True)
     reporting._authorize_sql_target("octopus")  # no raise
 
 
 def test_unknown_target_is_noop(monkeypatch):
     # Unknown targets are left for _run_sql to reject as a 400, not a 403 here.
-    monkeypatch.setattr(reporting, "has_permission", lambda p: False)
+    monkeypatch.setattr(_shared, "has_permission", lambda p: False)
     reporting._authorize_sql_target("bogus")  # no raise
